+++ CHANGELOG.md
@@ -5,3 +5,12 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 This Change Log format is suggested by
 <https://github.com/olivierlacan/keep-a-changelog/blob/master/CHANGELOG.md>
+
+<a name="0.2.9"></a>
+## 0.2.9 (2017-01-01)
+
+#### Added
+
+*   Install and configure gitguild
+*   Ledger for this guild
+*   isysd becomes member